+++ application/comit_node/src/swap_protocols/rfc003/alice/actions/non_erc20_actions.rs
@@ -1,7 +1,8 @@
 use crate::swap_protocols::{
     asset::Asset,
     rfc003::{
-        alice::{self, actions::CreateActions, SwapCommunication},
+        actions::CreateActions,
+        alice::{self, SwapCommunication},
         state_machine::HtlcParams,
         Action, Actions, Ledger, LedgerState,
     },
